runners/shared/requirements-pia.txt
pyyaml
numpy
torch
torchvision
